requirements.txt
PyPDF2
langchain
sentence-transformers
chromadb
gradio
python-dotenv
langchain_groq
langchain-community